Change script/fetch-html to not use ssh
script/fetch-html was ssh'ing into servers and running scripts on the
server side to generate a single html for a book and downloading it.

This is not ideal for the long term because it means that textbook devs
need to have ssh access to servers for them to download a book.

As part of the baked pdf CLI, we changed this to use "neb get" to
download the book with collection.xml and index.cnxml, which is all
public via https.  Then use "neb assemble" to create the single html.

```
./script/fetch-html biology
```

will download the book in `data/biology/raw` and the single html is
stored in `data/biology/collection.assembled.xhtml`.

```
./script/fetch-html --with-resources biology
```

will download the book as above but also the resources (e.g. images).