Change script/fetch-html to not use ssh (#983)
* Change cnx-recipes to use python 3.6.8

The baked-pdf CLI is in nebuchadnezzar which requires python >= 3.5.  To
prepare for changing the scripts to use neb, we need to change the
python version:

- Change `.python-version` which is used for installing the correct
  python version in pyenv;

- Change `Jenkinsfile` to use python 3 and create a universal wheel file
  so it's possible to install cnx-recipes in python 2 and python 3.
  Python 2 support is essential for cnx-deploy.

- Change `script/bootstrap` to check the virtualenv python version and
  recreate the virtualenv if necessary.

- Install cnx-easybake>=1.2.5 for python 3 compatibility fixes

* Change script/fetch-html to not use ssh

script/fetch-html was ssh'ing into servers and running scripts on the
server side to generate a single html for a book and downloading it.

This is not ideal for the long term because it means that textbook devs
need to have ssh access to servers for them to download a book.

As part of the baked pdf CLI, we changed this to use "neb get" to
download the book with collection.xml and index.cnxml, which is all
public via https.  Then use "neb assemble" to create the single html.

```
./script/fetch-html biology
```

will download the book in `data/biology/raw` and the single html is
stored in `data/biology/collection.assembled.xhtml`.

```
./script/fetch-html --with-resources biology
```

will download the book as above but also the resources (e.g. images).

* Change script/bake-book to use outputs from fetch-html

The fetch-html script now generates the book html in the format
`./data/{book_name}/collection.assembled.xhtml`.  Adjust the bake-book
script to use this as the input file and change the output to
`./data/{book_name}/collection.baked.xhtml`.

* Require the latest released version of Nebuchadnezzar

* Update the README to reflect changes to fetch-html

* Update scripts to use new data directory format

* format the intro-business snippet

since diffing was difficult and the content changed

* update references to -baked.xhtml in scripts and docs